FIX newIntent() for starting a blank intent
The newIntent method in ViewerActivity does not tell which activity to
start and which activity started the intent. Fixed it by requiring the
caller of this method to pass the Context of the activity that wants to
get the intent. The Intent object inside the method now adds the
context object and the class of ViewerActivity as parameters to the
returned Intent.