FIX call to EntryStash method of menu_new_entry
The menu_new_entry optionsMenu in JornalListFragment has been using the
old implemntation of entry stash where a new entry can be easily added
to the list of entries by getting a reference to the backing arraylist
of the EntryStash object and then appending the new entry to that list.
The Fixed code now uses the new implementation that  exposes the
addEntry method for adding an entry to the EntryStash.