FIX changing from pm to am when changing dates
The current time changes from pm to am when changing the date of the
Entry. This is because in the retrieval of the current hour of the
entry, Calendar.HOUR was used which always results to am time. The fix
for the problem is that the call to retrieve the current hour of the
Entry now uses the Calendar.HOUR_OF_DAY constant to get the correct
time in am/pm format.